+++ requests_magic/middleware.py
@@ -1,3 +1,7 @@
+"""
+下载中间件们
+"""
+
 from .exceptions import *
 from .utils import *
 import requests
@@ -5,7 +9,18 @@ import requests
 
 def requests_downloader(request) -> requests.Response:
     """
-    Default download function (downloader)
+    这是默认的下载中间件，基于 requests
+    Parameters
+    ----------
+    request
+        请求
+    Raises
+    ----------
+    RequestTimeoutError
+        请求超时，这会尝试重试请求
+    Returns
+    -------
+    下载的最终结果
     """
     kwargs = request_to_requests_kwargs(request)
     try:
@@ -16,11 +31,13 @@ def requests_downloader(request) -> requests.Response:
 
 def requests_downloader_filter(response: requests.Response, request) -> None:
     """
-    Default downloader filter
+    默认的请求过滤器，基于 requests，用来过滤下载完成后的结果，没有返回值，但可能抛出各种异常
+    Parameters
+    ----------
+    response
+        下载好的结果
+    request
+        请求
     """
     if response.status_code >= 400:
         raise RequestHttpError(request, response.status_code)
-
-
-def spider_parse_before(result, request):
-    pass