+++ requests_magic/scheduler.py
@@ -9,18 +9,31 @@ import time
 
 
 class Scheduler(threading.Thread):
+    """
+    调度器，核心组件，负责请求管理与 item 转发
+    """
 
-    def __init__(self, pipeline: Union[Pipeline, List[Pipeline]], max_link: int = 12,
+    def __init__(self, pipelines: Union[Pipeline, List[Pipeline]], max_link: int = 12,
                  request_interval: float = 0, distinct: bool = True):
         """
-        A request scheduler, which is the core of the crawler
+        调度器，核心组件，负责请求管理与 item 转发
+        Parameters
+        ----------
+        pipelines
+            管道们
+        max_link
+            最大连接数，默认：12
+        request_interval
+            请求间隔时间，默认：0秒
+        distinct
+            是否开启去重，默认开启
         """
         super().__init__()
         self.distinct = distinct
-        if isinstance(pipeline, list):
-            self.pipeline_list: List[Pipeline] = pipeline
-        elif isinstance(pipeline, Pipeline):
-            self.pipeline_list: List[Pipeline] = [pipeline]
+        if isinstance(pipelines, list):
+            self.pipeline_list: List[Pipeline] = pipelines
+        elif isinstance(pipelines, Pipeline):
+            self.pipeline_list: List[Pipeline] = [pipelines]
         else:
             raise SchedulerError('pipeline must be a Pipeline or Pipeline list')
         self.max_link: int = max_link
@@ -33,6 +46,14 @@ class Scheduler(threading.Thread):
                 pl.start()
 
     def add_request(self, request: Request) -> None:
+        """
+        添加一个新的请求（不会立刻执行）
+        Parameters
+        ----------
+        request
+            请求
+        """
+        Logger.debug(f"Add new request {request}, list requests len:{len(self.requests)}")
         md5: str = request.md5()
         if self.distinct and md5 in self.requests_md5:
             Logger.warning(f'Repeated request: {request}')
@@ -42,6 +63,16 @@ class Scheduler(threading.Thread):
         self.requests_md5.append(md5)
 
     def add_item(self, item: Item, from_spider=None) -> None:
+        """
+        添加一个新的 Item，这会转发给每个管道
+        Parameters
+        ----------
+        item
+            Item
+        from_spider
+            产生这个 Item 的爬虫，可以为空，空了也没啥问题
+        """
+        Logger.debug(f"Add new item {item}")
         item.scheduler = self
         item.spider = from_spider
         for pl in self.pipeline_list:
@@ -49,6 +80,15 @@ class Scheduler(threading.Thread):
                 pl.add_item(item)
 
     def add_callback_result(self, result_ite, from_spider=None) -> None:
+        """
+        自动解析解析函数返回的结果，会自动迭代、添加请求或转发 Item
+        Parameters
+        ----------
+        result_ite
+            结果
+        from_spider
+            产生结果的爬虫
+        """
         if result_ite is None:
             return
         if not isinstance(result_ite, Iterable):
@@ -59,12 +99,14 @@ class Scheduler(threading.Thread):
             elif isinstance(result, Item):
                 self.add_item(result, from_spider=from_spider)
             else:
-                Logger.warning(
+                Logger.error(
                     f"Cannot handle {str(type(result))}, " +
                     f"Please do not generate it in spider methods.")
 
     def run(self) -> None:
-        # Create link
+        """
+        开始处理请求队列
+        """
         while True:
             try:
                 if len(self.link_requests) < self.max_link and self.requests:
@@ -78,11 +120,15 @@ class Scheduler(threading.Thread):
             except Exception as e:
                 Logger.error(e)
 
-    def downloader_over(self, result, request: Request) -> None:
+    def downloader_finish(self, result, request: Request) -> None:
         """
-        Call from request. When the request is completed,
-        remove the request, execute the callback and process the result
+        当下载完成时，由 Request 调用。
+        这会在请求队列中移除这个请求并自动调用解析函数
         """
+        Logger.debug(f"Over a request {request}")
+        if request not in self.link_requests:
+            Logger.error(f"The completed download is not in link_request list. {request}")
+            return
         self.link_requests.remove(request)
         result = request.preparse(result, request)
         call = request.callback(result, request)
@@ -90,6 +136,23 @@ class Scheduler(threading.Thread):
 
     def downloader_abandon(self, request: Request) -> None:
         """
-        abandon a request
+        放弃一个请求
         """
+        Logger.debug(f"Abandon a request {request}")
+        if request not in self.link_requests:
+            Logger.error(f"The abandoned download is not in link_request list. {request}")
+            return
         self.link_requests.remove(request)
+
+    def downloader_retry(self, request: Request) -> None:
+        """
+        重试一个请求
+        """
+        if request not in self.link_requests:
+            Logger.error(f"The retry download is not in link_request list. {request}")
+            return
+        if request in self.requests:
+            Logger.error(f"The retry download have not started to request. {request}")
+            return
+        self.link_requests.remove(request)
+        self.requests.append(request)